Button "Refresh" must be active to give ability to reset changes (#61)
### Description of issue

Button `Refresh` must give ability to reset changes but it becomes
inactive when user makes changes. Need to remove logic to disable the
button and introduce confirmation dialog if there are unsaved changes

### Description of fix

- implemented new custom components
- `PopupDialog` - represents popup dialog which takes 3 parameters:
`Title`, `Message` and `Buttons`
- `PopupDialogHost` - container to show popup dialog within. It contains
method `ShowDialog` which creates new `PopupDialog` with provided
parameters and put it inside `PopupDialogHost` component. Waiting for
pressing buttons in dialogs was implemented via small hack to "freeze"
thread. Need to keep in mind that potentially it can harm performance
- implemented `PopupDialogHost` in main window
- added command to refresh notes to show confirmation dialog if there
are unsaved changes

### Screenshot 

![image](https://github.com/sergey-koryshev/idler/assets/11807074/e3e24550-34ac-4f35-b8d0-0d241ff2f11c)