Notes lost sharpness (#62)
### Description of issue

Notes related text boxes lost sharpness when they are not blurred due to
the fact that when `Radius` of `BlurEffect` equals to `0` then it seems
still blurred despite zero-value. I didn't found information related
this so probably it can be considered as WPF related bug

### Description of fix

I fixed it by removing `BlurEffect` for not blurred elements. It works
as expected but as far as I understand if we decide to implement some
animation for blur action we will have some problems with such approach